// src/utils/enums/error-response.enum.ts
export enum ErrorMessages {
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_EXISTS = 'USER_EXISTS',
  USER_NOT_ACTIVE = 'USER_NOT_ACTIVE',
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  INVALID_EMAIL = 'INVALID_EMAIL',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TOKEN_NOT_FOUND = 'TOKEN_NOT_FOUND',
  PASSWORD_DID_NOT_MATCHED = 'PASSWORD_DID_NOT_MATCHED',
  ACCESS_DENIED = 'ACCESS_DENIED',
  OTP_DID_NOT_MATCH = 'OTP_DID_NOT_MATCH',
  OTP_EXPIRED = 'OTP_EXPIRED',
  NOT_ALLOWED = 'NOT_ALLOWED',
  USERNAME_EXIST = 'USERNAME_EXIST',
  PHONE_NUMBER_EXIST = 'PHONE_NUMBER_EXIST',
  OTP_NOT_FOUND = 'OTP_NOT_FOUND',
  SESSION_EXPIRED = 'SESSION_EXPIRED',
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  AUTH_NOT_FOUND = 'AUTH_NOT_FOUND',
  TELEGRAM_AUTH_NOT_FOUND = 'TELEGRAM_AUTH_NOT_FOUND',
  CHANNEL_ALREADY_EXISTS = 'CHANNEL_ALREADY_EXISTS',
  CHANNEL_NOT_FOUND = 'CHANNEL_NOT_FOUND',
  CHANNEL_NOT_ACTIVE = 'CHANNEL_NOT_ACTIVE',
  USERNAME_ALREADY_EXISTS = 'USERNAME_ALREADY_EXISTS',
  ALREADY_SUBSCRIBED = 'ALREADY_SUBSCRIBED',
  NO_ENOUGH_MONEY = 'NO_ENOUGH_MONEY',
}