// src/utils/enums/success-response.enum.ts
export enum SuccessMessages {
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_DELETED = 'USER_DELETED',
  USER_FOUND = 'USER_FOUND',
  USER_LOGIN = 'USER_LOGIN',
  USER_FORGOT_PASSWORD = 'USER_FORGOT_PASSWORD',
  USER_PASSWORD_UPDATED = 'USER_PASSWORD_UPDATED',
  USER_TELEGRAM_LOGIN = 'USER_TELEGRAM_LOGIN',
  OTP_MATCHED = 'OTP_MATCHED',
  CHANNEL_UPDATED = 'CHANNEL_UPDATED',
  CHANNEL_DELETED = 'CHANNEL_DELETED',
}